patches: Exclude URI from tab name for named sites.
By default, FileZilla will show the full URI of a connection
in the tab name, even if it has a name in the Site Manager.
This adjusts the UI to show only the custom name if there is one,
falling back to showing the URI for "ephemeral" connections
that do not have the name.